--- src/shared/hooks/useDisclosure.ts
@@ -1,3 +1,0 @@
-import { useDisclosure as useReactHookDisclosure } from '@mantine/hooks';
-
-export const useDisclosure = useReactHookDisclosure;
