--- src/shared/hooks/useElementSize.ts
@@ -1,3 +1,0 @@
-import { useElementSize as ElementSize } from '@mantine/hooks';
-
-export const useElementSize = ElementSize;
