--- src/shared/hooks/useMediaQuery.ts
@@ -1,3 +1,0 @@
-import { useMediaQuery as MediaQuery } from '@mantine/hooks';
-
-export const useMediaQuery = MediaQuery;
