+++ src/shared/lib/mantine/hooks.ts
@@ -1,1 +1,2 @@
 export { useMantineTheme } from '@mantine/core';
+export { useMediaQuery, useDisclosure } from '@mantine/hooks';
