+++ src/shared/lib/mantine/index.ts
@@ -1,3 +1,3 @@
 export { colors } from './colors';
 export { ThemeProvider } from './ThemeProvider';
-export { useMantineTheme } from './hooks';
+export { useMantineTheme, useMediaQuery, useDisclosure } from './hooks';
